Feat[controllers]: implements controllers and tests e2e

+++ src/use-cases/users/delete-user.ts
@@ -1,4 +1,5 @@
 import { UsersRepository } from '@/repositories/users-repository'
+import { UserNotExistsError } from '../errors/user-not-exists'
 
 interface DeleteUserUseCaseRequest {
   id: string
@@ -14,6 +15,12 @@ export class DeleteUser {
   async execute({
     id,
   }: DeleteUserUseCaseRequest): Promise<DeleteUserUseCaseResponse> {
+    const user = await this.usersRepository.findById(id)
+
+    if (!user) {
+      throw new UserNotExistsError()
+    }
+
     const { message } = await this.usersRepository.delete(id)
 
     return { message }
